backend/requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
celery[redis]==5.4.0
django-environ==0.11.2
psycopg[binary]==3.2.3
redis==5.2.1
reportlab==4.2.5
gunicorn==23.0.0
